feat: returning long as deleted/updated count

+++ src/main/java/com/github/sepgh/testudo/operation/CollectionDeleteOperation.java
@@ -5,5 +5,5 @@ import com.github.sepgh.testudo.operation.query.Query;
 public interface CollectionDeleteOperation<T extends Number & Comparable<T>> {
     CollectionDeleteOperation<T> query(Query query);
     Query getQuery();
-    int execute();
+    long execute();
 }
